image-processing-service/requirements.txt
argcomplete==2.0.0
attrs==22.1.0
bcrypt==3.2.2
Beaker==1.10.0
beautifulsoup4==4.11.0
blivet==3.5.0
blivet-gui==2.4.0
Brlapi==0.8.4
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
cryptography==37.0.2
cupshelpers==1.0
dasbus==1.6
dbus-python==1.3.2
decorator==5.1.1
deepmerge==1.1.0
distro==1.7.0
dnspython==2.2.1
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
dotmap==1.3.30
fedora-third-party==0.10
file-magic==0.4.0
Flask==2.2.2
fros==1.1
gpg==1.17.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
humanize==3.13.1
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.10.0
jupyter-core==4.10.0
langtable==0.0.61
libcomps==0.1.18
lxml==4.9.1
Mako==1.1.4
MarkupSafe==2.1.1
nftables==0.1
numpy==1.22.0
olefile==0.46
packaging==21.3
paramiko==2.12.0
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==9.2.0
ply==3.11
priority==2.0.0
productmd==1.33
ptyprocess==0.6.0
pwquality==1.4.4
pyasn1==0.4.8
pycairo==1.21.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.1
pyenchant==3.2.2
Pygments==2.14.0
PyGObject==3.42.2
pykickstart==3.41
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyparted==3.12.0
pyrsistent==0.18.1
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-meh==0.50
pyudev==0.23.2
pyxdg==0.27
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-ftp==0.3.1
rich==13.1.0
rpm==4.18.0
selinux==3.4
sepolicy==3.4
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.4
soupsieve==2.3.2.post1
systemd-python==235
Tempita==0.5.2
texttable==1.6.4
toml==0.10.2
traitlets==5.2.1.post0
urllib3==1.26.12
websocket-client==1.3.3
Werkzeug==2.2.2
wsproto==1.2.0
